SVG images double rendering on Android
Fixing RNSVGImageShadowNode to no longer dangerously store a reference
to a DataSource result, since it's only valid for the lifetime of the
onNewResultImpl() method. Instead, the images are rendered from the
Fresco image bitmap cache.

Updating RNSVGVirtualNode to call getScreenDisplayMetrics() instead of
getWindowDisplayMetrics(), since the latter is deprecated.

Synchronizing access to RNSVGSvgViewShadowNode.drawChildren(), since
it can be called by multiple threads simultaneously (the native UI
thread and the asynchronous Image loader callback).