Fix setNativeProps type
Setting Object to the props type does not allow passing any object to the method. So in TypeScript when using the method for example with Path, you will get an error when passing the d:
this.component?.setNativeProps({ d }) => 
Argument of type '{ d: string; }' is not assignable to parameter of type 'Object & { matrix?: ColumnMajorTransformMatrix | undefined; fill?: ColorValue | undefined; } & TransformProps'.
  Object literal may only specify known properties, and 'd' does not exist in type 'Object & { matrix?: ColumnMajorTransformMatrix | undefined; fill?: ColorValue | undefined; } & TransformProps'.ts(2345)

Instead ob Object, you should use at least P so it allows the original props of the component you are using.