Minor fix to prevent crash when Loaded fires twice
There are many reasons `Loaded` can fire twice. Usually, it will happen
when the native SvgView component is moved, but it can also happen if an
ancestor of SvgView is reparented. There is a crash that appears to
occur when we attempt to call CreateResources twice. This resolves that
crash.