Fix crash when SVG is unloaded by XAML
XAML may unload the root SVG panel, which will trigger a recursive
unload. For GroupView, this unload will clear the child collection.
However, active operations on the SVG from the React Native UIManager
may still be in flight, in which case we may attempt to remove a child
that has already been cleared by the root unload.