chore: Remove legacy version check in `react-native.config.js` (#2503)
## Description

Removes legacy version check from `react-native.config.js`. 

## Changes

The community CLI is no longer a dependency for React Native, meaning it
may not be installed in every project. For example, projects that use
Expo won't include the `@react-native-community/cli-*` packages.
Consequently, this check will fail, leading to issues with autolinking
components.

This version check can be removed safely, as the 9th version of the CLI
was released two years ago.

## Tests

This change was tested using `expo` project on react native `0.76`.

Co-authored-by: Jakub Grzywacz <jakub.grzywacz@swmansion.com>