fix(windows): svgs being rendered blurry when at high dpi (#2640)
# Summary
The windows svg rendering pass was always rendering the svg to a bitmap
the size of the svg at 100% display scale factor, which would then be
stretched at higher scale factors. This meant that the svg would be
blurry when running in scale factor systems.

With this change the svg will be rendered taking into account the scale
factor.

I also removed an invalid color="none" property set, which would cause
errors when running against a debug D2D instance. - That property had no
effect, as it was an invalid property value. -- Instead we just leave
the property unset when there is no color specified.

## Test Plan

Before:
![Screenshot 2025-02-27
110318](https://github.com/user-attachments/assets/aae0cb81-96a3-415e-a908-aa4e46c82a37)

After:
![Screenshot 2025-02-27
151859](https://github.com/user-attachments/assets/1ec86957-4a98-41f3-9ba3-4b400c7849f8)

## Checklist

<!-- Check completed item, when applicable, via: [X] -->

- [x] I have tested this on a device and a simulator
- [ ] I added documentation in `README.md`
- [ ] I updated the typed files (typescript)
- [ ] I added a test for the API in the `__tests__` folder