Remove explicit value for `usesMapBufferForStateData` template parameter
This wasn't present in older versions of React Native, and is removed from newer versions. It was kind of meant to be hidden, with the default value used, but was fully public. Let's rely on the default `false` instead of specifying an explicit value, for compatibility of versions of RN without it.