fix: handle error when defining a nil reference (#2717)
# Summary

That PR helped to solve issue #2715 

This issue arises on the iOS platform because
`NSMutableDictionary<NSString *, RNSVGNode *>` does not allow nil keys
or values. On Android, however, the `Map<String, VirtualView>`
implementation does allow null keys, which prevents this issue from
occurring there.

To address the iOS issue, we handle nil values for elements such as
`clipPaths, templates, painters, markers, masks, and filters`. Managing
`nil` values in `RNSVGSvgView.mm` is practical because the `define*`
method is used in multiple locations.

---------

Co-authored-by: Jakub Grzywacz <kontakt@jakubgrzywacz.pl>
Co-authored-by: Jakub Grzywacz <jakub.grzywacz@swmansion.com>