feat: add onLoad call in SvgWithCssUri component (#2751)
# Summary

This PR fixes: #2735 
 
Previously, the `SvgWithCssUri` component exposed an onLoad prop, but it
was never invoked. This PR ensures that onLoad is called once the remote
SVG has been successfully fetched.

## Test Plan
Pass `onLoad` function to `SvgWithCssUri` component and confirm that the
onLoad callback is called once the SVG has successfully loaded.

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ❌      |
| MacOS   |    ❌      |
| Android |    ❌      |
| Web     |    ✅      |