fix: pointerEvents handling on iOS (#2787)
# Summary
Fixes: #2784 
Fixes: #2690
Fixes: #2639

Rename `hitTest` to `betterHitTest` for new architecture to handle
properly `PointerEvents` in SVG.
Before that change we override `hitTest` from
[RCTViewComponentView.mm](https://github.com/facebook/react-native/blob/3766e58b9aa50021bff508e215dcc18efb921435/packages/react-native/React/Fabric/Mounting/ComponentViews/View/RCTViewComponentView.mm#L673)
which cause incorrect handling of `pointerEvents`.
With this change we override `betterHitTest` that
`RCTViewComponentView.mm` use in `hitTest` based on `pointerEvents`
which fix the issue connected to pointerEvents.

## Test Plan

Run example from issue: #2784 

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅      |
| MacOS   |    ✅      |
| Android |    ❌      |
| Web     |    ❌      |