fix: positioning element inside viewbox (#2758)
# Summary
Fixes: #2078 
Fixes: #1343 

When `preserveAspectRatio !== "none"`, the initial `translateX` and
`translateY` values should be calculated after adjusting `scaleX` and
`scaleY` with `Math.min` or `Math.max` (depending on the `meetOrSlice`
value).
Previously `translateX` and `translateY` were computed before applying
this adjustment using the raw `scaleX` and `scaleY`. This caused
incorrect positioning of the content.

## Test Plan
Run examples from issues: #2078 and #1343

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅      |
| MacOS   |    ✅      |
| Android |    ✅      |
| Web     |    ❌      |