fix: set absolute fill position in shadow nodes to fix onPress hit test in js (#2840)
# Summary

Sets all SVG elements position to absolute fill instead of `{0, 0, 0,
0}` in their shadow nodes to return `Svg` element layout metrics in
measure function to fix `onPress` handling in js.

A limitation of this approach is that touch cancellation by dragging the
finger outside the element is not supported - touch will only cancel if
moved outside the entire SVG element.

## Test Plan

Test `FabricExample`